refactor: extract resolve and load functions to utils.js
